Add gnmi_dump tool for debug and unit test (#60)
Why I did it
Add tool for debug and unit test.

How I did it
Use share memory to support counters.
Move check for EnableTranslibWrite to beginning of Set(). This change will simplify the logic for gnmi set fail, and no need to authenticate when EnableTranslibWrite is false.

How to verify it
Build image and run unit test.