Support multiple database for DPU (#188)
#### Why I did it
We have updated database configuration to support multiple databases for smartswitch.
And then we need to update sonic-gnmi to support new database configuration and new GNMI path for smartswitch.
HLD PR: https://github.com/sonic-net/SONiC/pull/1625

#### How I did it
* Support new database configuration

```
{
    "INCLUDES" : [
        {
            "include" : "../../redis/sonic-db/database_config.json"
        },
        {
            "container_name" : "dpu0",
            "include" : "../../redisdpu0/sonic-db/database_config.json"
        }
    ],
    "VERSION" : "1.0"
}
```

Use a single layer instance for each device to support multiple asic and multiple dpu.
For example, for multiple asic, the instances can be named localhost, asic0, asic1, and so on. For multiple dpu, the instances can be named localhost, dpu0, dpu1, and so on. For multiple asic npu and multiple dpu, the instances can be named localhost, asic0, asic1, ..., dpu0, dpu1, and so on.

```
        "DPU_APPL_DB" : {
            "id" : 15,
            "separator": ":",
            "instance" : "redis",
            "format": "proto"
        },
```

* Update GNMI path to support multiple database

I add a new element in GNMI path to support database instance.
```
/APPL_DB/dpu0/DASH_VNET_TABLE/vnet1/
```
The second element is used for target database instance, and it can be localhost, asic0, asic1, ..., dpu0, dpu1, and so on.

#### How to verify it
Run unit test for sonic-gnmi.
Run end to end test for dash and gnmi.