Use the latest libswsscommon artifacts from the repo
For libswsscommon, instead of getting the artifacts from the
sonic-buildimage repo, get it from the sonic-swss-common repo instead.
The version in sonic-buildimage might be behind because the submodule
update hasn't happened. That way, it's testing with the latest code,
and also matches the go compilation where it uses the latest of
sonic-swss-common.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>