Issue 357: Use ConfigReplace when full config is sent over gNMI
When a full config is sent of gNMI, the current code saves it in a file
and will restore it when a gNOI reboot is requested (similar to what
'config reload' would do). This is an expensive operation, as all
services must be brought down before the config can be applied. Instead,
trigger a call to 'config replace' so that the config can be applied
immediately without needed to restart services. There is already a D-Bus
endpoint for 'config replace', so just add a D-Bus gNMI client handler
and then invoke that handler when passed a full config.

A unit test of the new handler was added. Testcases for the modified SET
handling logic were added.