Issue #22362: Call ConfigReload on full config gnmi update
The full config update handler already saves the config to a file and
does yang validation on it. The previous implementation relied on the
Reboot command to trigger a ConfigReload, but that command has been
implemented on the system level and no longer performs a reload, leaving
this operation non-functional. However, we can call ConfigReload
directly from the handler and bypass the Reboot, restoring this
operation.

Modify unit tests to expect test D-Bus calls.