Add support for ACL counter
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

##### Work item tracking
- Microsoft ADO **(number only)**:32514092

#### Why I did it
The dataplane ACL is utilized in AI frontend devices to block hosts from accessing unauthorized services. To enhance the visibility of each ACL rule's functionality and assist with troubleshooting, it is necessary to obtain telemetry data for the counters of each ACL rule through streaming telemetry.

#### How I did it
Implement VtoR Path Mapping for COUNTERS_DB queries towards ACL counters
Multi-asic is NOT supported for this feature currently, which means same rule under multiple namespace is not well organised in the map yet.

#### How to verify it
<img width="2531" height="1222" alt="image" src="https://github.com/user-attachments/assets/bf5f300d-55e3-492b-89fa-8d821814dea0" />

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 201811
- [ ] 201911
- [ ] 202006
- [ ] 202012
- [ ] 202106
- [ ] 202111

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/SONiC/wiki/Configuration.
-->

#### A picture of a cute animal (not mandatory but encouraged)