fix(pyang): Add build tag to generated main.go to allow testing
Adds the '//go:build !test' build constraint to the pyang client
generator template (main.j2).

This prevents the flag definitions in the generated 'main' packages
from being included in test builds, which resolves the 'flag redefined'
panic when running 'go test ./...'. This change unblocks the
consolidation of the test commands in the Makefile.

Signed-off-by: Dawei Huang <daweihuang@microsoft.com>