fix: use explicit GOPATH for gotestsum in memory leak tests
Use $(shell $(GO) env GOPATH)/bin/gotestsum to ensure the correct
path is used when running with sudo. Normal user has GOPATH=/tmp/go
but sudo has GOPATH=/root/go, so we need the explicit path to find
the gotestsum binary installed by the normal user.

This matches the pattern we tested successfully in the container.

Signed-off-by: Dawei Huang <daweihuang@microsoft.com>